Refactor to use fix_times everywhere
