+++ src/pytroll_watchers/publisher.py
@@ -2,9 +2,11 @@
 
 from contextlib import closing, suppress
 from copy import deepcopy
+import datetime
 
 from posttroll.message import Message
 from posttroll.publisher import create_publisher_from_dict_config
+from trollsift import parse
 
 
 def file_publisher_from_generator(generator, publisher_config, message_config):
@@ -29,3 +31,29 @@ def file_publisher_from_generator(generator, publisher_config, message_config):
             amended_message_config["data"].update(file_metadata)
             msg = Message(**amended_message_config)
             publisher.send(str(msg))
+
+
+def fix_times(info):
+    """Fix times so that date and time components are combined, and start time is before end time."""
+    if "start_date" in info:
+        info["start_time"] = datetime.datetime.combine(info["start_date"].date(),
+                                                       info["start_time"].time())
+        if "end_date" not in info:
+            info["end_date"] = info["start_date"]
+        del info["start_date"]
+    if "end_date" in info:
+        info["end_time"] = datetime.datetime.combine(info["end_date"].date(),
+                                                     info["end_time"].time())
+        del info["end_date"]
+    if "end_time" in info:
+        while info["start_time"] > info["end_time"]:
+            info["end_time"] += datetime.timedelta(days=1)
+
+def parse_metadata(file_pattern, path):
+    """Parse metadata from the filename."""
+    if file_pattern is not None:
+        file_metadata = parse(file_pattern, path)
+        fix_times(file_metadata)
+    else:
+        file_metadata = {}
+    return file_metadata
